Fixed plotly figure design with help of Natashia, separated frequency plot to its own component with props, will have to change it a bit to work with populations and well as superpopulations
